Avoid some warnings from gcc.
Assume the indentation expresses correctly how the statements are meant to be
guarded by the ifs.

cc -std=gnu99 -Wall -Wshadow -Wextra -pedantic -Woverflow -Wstrict-overflow -ggd
b  -O3  -c -o assemble.o assemble.c
assemble.c: In function 'assemble':
assemble.c:963:29: warning: this 'if' clause does not guard... [-Wmisleading-ind
entation]

  963 |                             if (op->value == P_PRINT)
      |                             ^~
assemble.c:966:33: note: ...this statement, but the latter is misleadingly inden
ted as if it were guarded by the 'if'
  966 |                                 if (show_print_lines) {
      |                                 ^~

This last one I'm guessing the `if (show_print_lines)` is meant to be guarded
by the `if (op->value == P_PRINT)` because of the comment.